Rename getXXX and setXXX to just get and set

+++ valen-game-darkages/src/main/java/be/twofold/valen/game/darkages/reader/resources/Resources.java
@@ -28,7 +28,7 @@ public record Resources(
         // assert channel.position() == header.addrPathStringOffsets();
         var numStrings = reader.readLongAsInt();
         var offsets = reader.readLongsAsInts(numStrings);
-        var stringBufferLength = Math.toIntExact(header.addrDependencyEntries() - header.addrPathStringOffsets() - (numStrings + 1) * (long) Long.BYTES);
+        var stringBufferLength = header.addrDependencyEntries() - header.addrPathStringOffsets() - (numStrings + 1) * Long.BYTES;
         var stringBufferRaw = reader.readBytesStruct(stringBufferLength);
         var stringBuffer = DECODER.decode(stringBufferRaw.asBuffer()).toString();
         var pathStrings = Arrays.stream(offsets)
